requirements.txt
django>=3.0,<4.0
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==20.1.0